整理代码
