+++ src/TelescopeSourceGenerator/Analyzers/TelescopeIncrementalGenerator.cs
@@ -4,7 +4,7 @@ using System.Collections.Immutable;
 using System.Diagnostics;
 using System.Linq;
 using System.Threading;
-
+using dotnetCampus.Telescope.SourceGeneratorAnalyzers.Core;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
